feat(pvmt): handle refine testing

+++ crates/core/src/vm/refine.rs
@@ -7,12 +7,15 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone, Default)]
 pub struct RefineParams {
     /// (c) the core index
+    #[serde(with = "codec::compact")]
     pub core: u16,
 
     /// (i) the work item index
-    pub index: usize,
+    #[serde(with = "codec::compact")]
+    pub index: u16,
 
     /// (w_s) the service id
+    #[serde(with = "codec::compact")]
     pub id: ServiceId,
 
     /// (y) the payload
